Finish aligning Scala-CLI and mill
All commands work from readme files
