+++ backend/mill/build.mill
@@ -1,0 +1,19 @@
+//| mill-version: 1.0.3
+
+package build
+
+import mill.*, scalalib.*
+
+object minimumViableProduct extends ScalaModule {
+  def scalaVersion = "3.7.2"
+  def mvnDeps = Seq(
+    mvn"com.lihaoyi::cask:0.10.2",
+    mvn"com.lihaoyi::upickle:4.2.1",
+    mvn"io.github.kitlangton::neotype:0.3.25"
+  )
+
+  object test extends ScalaTests {
+    def mvnDeps = Seq(mvn"com.lihaoyi::utest:0.9.0")
+    def testFramework = "utest.runner.Framework"
+  }
+}
