+++ backend/scalaCLI/proofOfConcept/A_Invariants.scala
