+++ backend/scalaCLI/proofOfConcept/B_Rules.scala
